♻️ [Refactoring]: #455 Phase 4 escape カテゴリを tests/escape.rs に分離
単純エスケープ 8 種（n / r / t / b / f / ( / ) / バックスラッシュ）の
テスト 8 件を literal_string/tests/escape.rs に切り出し、tests.rs ハブに
`mod escape;` 宣言を追加。
